data-pipeline/requirements.txt
google-cloud-bigquery==3.41.0
google-cloud-bigquery-storage==2.37.0
db-dtypes==1.5.1
polars==1.39.3
pyarrow==23.0.1
pandas==2.3.3
python-dotenv==1.2.2
